Switch from mcapi.us to minecraft-server-util
HTTPS requests to mcapi.us started returning "error code: 1020" instead
of valid JSON, causing the ping() function for Minecraft to fail every
time it was called. The error code translates to "Access Denied" and is
mainly used by Cloudflare, this suggests that something has changed in
the Cloudflare configuration for MCApi which has caused it to block all
of the bot's requests.

Since sending requests to an external website has proven to be
unreliable, I have decided to switch to using a node.js package called
minecraft-server-util. Adding this package will greatly simplify the
code for getting the status of Minecraft servers and will make it easy
to add support for Minecraft: Bedrock servers later down the line.

The switch means that the bot will now be communicating with Minecraft
servers directly to get their status rather than relying an external
website to handle that process. This will require that the bot can
reach each server on it's own using the hostname provided to it, which
could cause problems if the bot is hosted on the same network as the
Minecraft server and the server needs to be accessible over the
internet. However, this does open up the possibly of obtaining the
status of a Minecraft server that is only accessible on the local
network.

[Fixes #7]